Introduce color generators
