+++ lib/geo_pattern/color_generators/simple_generator.rb
@@ -1,0 +1,28 @@
+module GeoPattern
+  # Color generators
+  module ColorGenerators
+    # Simple one
+    class SimpleGenerator
+
+      private
+
+      attr_reader :color, :creator
+
+      public
+
+      # New
+      #
+      # @param [String] color
+      #   HTML color string, #0a0a0a
+      def initialize(color, creator = Color)
+        @color = color
+        @creator = creator
+      end
+
+      # Generator color
+      def generate
+        Color.new(color)
+      end
+    end
+  end
+end
